scheduler/python/requirements.txt
google-api-python-client==2.131.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
python-dateutil==2.9.0.post0
